backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
transformers>=4.30.0
spandrel>=0.3.0
diffusers>=0.27.0
accelerate>=0.27.0
optimum[intel]
fastapi
uvicorn